v2.0.0
Hello everyone, just wanted to put in a little update to the RPC to make it a lot more cleaner. As a coder, the first version of a project is never clean enough. I'm a lot more prouder of this version so enjoy :)

Updates:
 - Detect when a livestream updates their information e.g. Category and Title
 - Change the Client ID to be custom so instead of it saying "Live on Kick" you can change it to something else instead
 - Cleaned up the logging system, only time you will get a "spam" of messages in the console is when my KickAPI is gathering information from the kick endpoint. I'm not sure why it's not silenced as I silenced it in the actual repo but oh well

Overall I'm pretty happy with this update. I hope you all like it :)